feat: adjust with latest jina clean up test
